Initial Commit
In this initial commit I have done the groundwork for the actual event store
and event sourcing layer. More testing needs to be done if all scenario's are covered.
We are doing this by implementing a test project in the examples folder that shows
how to implement a blog using event sourcing.

In future commits we will also add ACL, snapshotting and optimistic locking
heuristics to the event store packages.

Please note that I have split the library into an src and lib folder; the src folder is the
main event store layer (which knows nothing about aggregates, etc) and each folder
in the lib folder is a separate layer / package that can wrap the event store to add
additional functionality.