Removed meta data from Streams
The concept of meta data on a stream was an interesting concept but could
probably be better modeled in a different way. The only use-case that I currently have for
that is ACL and it would be better to track the ACL using an event-sourced representation
that can be queried.

By using an event-sourced representation we eat our own dog food and provide a flexible
mechanism that is capable of tracking changes in the ACL.